Refactors document list into standalone component
